Workaround for floating point error
Implemented a new function meant to be called before costly Vector2 operations. Its purpose is to give back a set of vectors that are simply moved as close to Vector2.ZERO as possible, which the costly operations can be performed on, before being re-adjusted to the same relative positions.

This should, hopefully, squash a good chunk of floating point error related problems without the users needing to compile a double precision version of GODOT.

Additionally, implemented a faster pair of loops for `_make_outer_polygon`, making it so every pair of edges is only compared once, rather than twice as it was previously.